Beta almost finished

+++ src/signal.c
@@ -9,50 +9,42 @@
 #include "./change.h"
 #include "./regfile.h"
 
-/**
- * 1º-> funções de reigisto de sinais
- * 2ª-> aplicar os signhandlers no xmod.c
- * 3ª-> aplicar o sinal de fim de fil de processo filho
- * 
- */
+int signSIGINTAnwserHandler(char *answer) {
+    if (answer == NULL) return -1;
 
-int signSIGINTAnwserHandler(char answer[]) {
-    //Se for igual a 1, vai mandar todos os procesos abortarem
-    if(answer[0] =='1') {
-        for(int i = 0; i<numberOfChildren; i++){
-            kill(childProcesses[i], SIGUSR1);
+    int status;
+
+    switch (answer[0]) {
+    case '1':
+        for(int i = 0; i < numberOfChildren; i++) {
             eventSignalSent(SIGUSR1, childProcesses[i]);
-        }
-        //Espera que os filhos todos morram antes de morrer
-        for(int i = 0; i<numberOfChildren; i++){
-            int status;
-            waitpid(childProcesses[i], &status,1 );
+            kill(childProcesses[i], SIGUSR1);
         }
 
-        endLogFile();
+        // Waits for the children to terminate before it terminates itself
+        for(int i = 0; i<numberOfChildren; i++)
+            waitpid(childProcesses[i], &status, 1);
+
         write(STDOUT_FILENO, "The program has been cancelled!\n", 33);
-
-        //Levanta o kill e regista tudo
-        eventSignalSent(SIGKILL, getpid());   //Envio que vou me matar
-        eventSignalRecv(SIGKILL);     //Recebo que me vou matar
-        eventProcExit(0);    //Aviso que me matei
-        raise(SIGKILL);     //Mato-me
-        return 0; 
-
-    } else if(answer[0] == '2') {
-        for(int i = 0; i<numberOfChildren; i++){
+        
+        // Raises kill and registers it
+        eventSignalSent(SIGKILL, getpid());
+        eventSignalRecv(SIGKILL);
+        eventProcExit(0);
+        raise(SIGKILL);
+        
+    case '2':
+        for(int i = 0; i < numberOfChildren; i++) {
             eventSignalSent(SIGCONT, childProcesses[i]);
             kill(childProcesses[i], SIGCONT);
         }
-
+        
         write(STDOUT_FILENO, "The program has been continued!\n", 33);
 
         return 0;
-
-    } else { return -1; }
+    }
+    return -1;
 }
-
-
 
 void signSIGINTHandler(int signo) {
     if (filename == NULL) {
@@ -64,7 +56,7 @@ void signSIGINTHandler(int signo) {
     eventSignalRecv(SIGINT);
 
     // Writing the pid; directory_name; number_of_files_found; number_of_files_modified
-    char buffer[100];
+    char buffer[50];
     snprintf(buffer, sizeof(buffer), "\n%d; %s; %d; %d", getpid(), filename, nftot, nfmod);
     write(STDOUT_FILENO, buffer, strlen(buffer));
 
@@ -79,43 +71,46 @@ void signSIGINTHandler(int signo) {
             write(STDOUT_FILENO,"Please insert '1' to stop or '2' to continue! ", 47);
             read(STDIN_FILENO, input, 2);
         }
-    } else {
+    }
+    else {
         // Children pause and registers the event
         eventSignalSent(SIGSTOP, getpid());
         eventSignalRecv(SIGSTOP);
         pause();
-    }
+    }   
 }
 
 
-void sighandlerSIGCONT(int signo){
-    //Regist que recebeu o sinal
+void sighandlerSIGCONT(int signo) {
+    if (filename == NULL) return;
+
+    // Registers the signal
     eventSignalRecv(SIGCONT);
 
+    // Sends the signals to the children and registers it
     for (int i = 0; i < numberOfChildren; i++) {
-        //envia o sigcont aos filhos e regista
-        kill(childProcesses[i], SIGCONT);
         eventSignalSent(SIGCONT, childProcesses[i]);
+        kill(childProcesses[i], SIGCONT);
     }
-
 }
 
-void sigHandlerSIGUSR1(int signo){
+void sigHandlerSIGUSR1(int signo) {
+    if (filename == NULL) raise(SIGKILL);
+    
+    // Registers the signal
     eventSignalRecv(SIGUSR1);
 
-    for(int i = 0; i<numberOfChildren; i++){
-        kill(childProcesses[i], SIGUSR1);
+    // Sends the signals to the children and registers it
+    for (int i = 0; i < numberOfChildren; i++) {
         eventSignalSent(SIGUSR1, childProcesses[i]);
+        kill(childProcesses[i], SIGUSR1);
     }
 
     eventSignalSent(SIGKILL, getpid());
     eventSignalRecv(SIGKILL);
-    eventSignalSent(SIGCHLD, getppid());    //Aviso o paizinho que me vou matar
+    eventSignalSent(SIGCHLD, getppid());
     eventProcExit(0);
     raise(SIGKILL);
 }
 
-void sigHandlerSIGCHILD(int signo){
-    eventSignalRecv(SIGCHLD);
-}
-
+void sigHandlerSIGCHILD(int signo) {eventSignalRecv(SIGCHLD);}
